Fix clippy CI test, fix Tessellator clippy issues
